Python 3 compatibility fixes.
SIP and MGW messages are encoded/decoded.

Dialog.remote_sdp_session_version was compared before set.

Divisions are rounded to integers where necessary.

dict.keys() is converted to list if the dict is altered during iteration.

+++ g711.py
@@ -2163,7 +2163,7 @@ def decode_pcma_sample(u8):
 
 
 def encode_pcmu(x):
-    n = len(x) / 2
+    n = int(len(x) / 2)
     y = bytearray(n)
     
     for i in range(n):
@@ -2173,7 +2173,7 @@ def encode_pcmu(x):
 
 
 def encode_pcma(x):
-    n = len(x) / 2
+    n = int(len(x) / 2)
     y = bytearray(n)
     
     for i in range(n):
